pyproject.toml
[tool.poetry]
name = "chia_liquidity_provider"
version = "0.2.0"
description = "Hybrid liquidity farming with Chia offers"
authors = ["Julie Bettens <no-reply@bbjubjub,fr>"]
license = "GPL3plus"

[tool.poetry.dependencies]
python = "^3.9"
chia-blockchain = "^1.6.2"
aiohttp = "^3.8"
click = ">=7"
aiomisc = "^16.2.10"
aiosqlite = "^0.17.0"
xdg = "^5.1.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
mypy = "^0.971"
chia-dev-tools = "^1.1.3"

[[tool.poetry.packages]]
include = "chia_liquidity_provider"
from = "src"

[tool.poetry.scripts]
clp = "chia_liquidity_provider:main"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"

[tool.black]
line-length = 100

[tool.isort]
line_length = 100

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
